requirements.txt
langchain-core==0.3.78
langchain-community==0.3.31
langchain-google-genai==2.1.12
langchain-openai==0.3.35
langgraph==0.6.8
python-dotenv==1.0.1
streamlit==1.39.0
gradio==4.28.3